requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
Flask-Script
Flask-Login
psycopg2
PyJWT==2.0.1